Rename Cloud WorkspacesService to CloudWorkspacesServices to avoid confusion with WorkspaceService Update functions in CloudWorkspacesServices to include the word Cloud
